agents/requirements.txt
httpx>=0.27
python-dotenv>=1.0
requests>=2.31
matplotlib>=3.7
pandas>=2.0
# Install llmproxy separately:
#   pip install -e ../LLMProxy-main/py
